+++ demo/changelog-main.ts
@@ -1,0 +1,5 @@
+import { createApp } from "vue";
+import "./style.css";
+import ChangelogPage from "./ChangelogPage.vue";
+
+createApp(ChangelogPage).mount("#app");